requirements.txt
absl-py==2.0.0
astunparse==1.6.3
Authlib==1.3.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
Flask==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
gast==0.5.4
google-auth== 2.28.2
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.59.3
h5py==3.10.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
keras==2.15.0
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
ml-dtypes==0.2.0
nltk==3.8.1
numpy==1.26.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pillow==10.2.0
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
python-dotenv==1.0.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.28
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow== 2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
textblob==0.17.1
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.14.1
gunicorn==20.0.4
google-cloud-storage==2.16.0